# docs/swagger.yaml
definitions:
  controller.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controller.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controller.ageRatingCategoryInput:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  controller.movieInput:
    properties:
      age_rating_category_id:
        type: integer
      title:
        type: string
      year:
        type: integer
    type: object
  models.AgeRatingCategory:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Movie:
    properties:
      age_rating_category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /age-rating-categories:
    get:
      description: Get a list of AgeRatingCategory.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AgeRatingCategory'
            type: array
      security:
      - BearerToken: []
      summary: Get all AgeRatingCategory.
      tags:
      - AgeRatingCategory
    post:
      description: Creating a new AgeRatingCategory.
      parameters:
      - description: the body to create a new AgeRatingCategory
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.ageRatingCategoryInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AgeRatingCategory'
      security:
      - BearerToken: []
      summary: Create New AgeRatingCategory.
      tags:
      - AgeRatingCategory
  /age-rating-categories/{id}:
    delete:
      description: Delete a AgeRatingCategory by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: AgeRatingCategory id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one AgeRatingCategory.
      tags:
      - AgeRatingCategory
    patch:
      description: Update AgeRatingCategory by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: AgeRatingCategory id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age rating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.ageRatingCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AgeRatingCategory'
      security:
      - BearerToken: []
      summary: Update AgeRatingCategory.
      tags:
      - AgeRatingCategory
  /age-rating-categories/{id}/movies:
    get:
      description: Get all Movies by AgeRatingCategoryId.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: AgeRatingCategory id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
      security:
      - BearerToken: []
      summary: Get Movies.
      tags:
      - AgeRatingCategory
  /login:
    post:
      description: Logging in to get jwt token to access admin or user api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as as user.
      tags:
      - Auth
  /movie/{id}:
    delete:
      description: Delete a movie by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: movie id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one movie.
      tags:
      - Movie
  /movies:
    get:
      description: Get a list of Movies.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
      security:
      - BearerToken: []
      summary: Get all movies.
      tags:
      - Movie
    post:
      description: Creating a new Movie.
      parameters:
      - description: the body to create a new movie
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.movieInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
      security:
      - BearerToken: []
      summary: Create New Movie.
      tags:
      - Movie
  /movies/{id}:
    get:
      description: Get a Movie by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: movie id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
      security:
      - BearerToken: []
      summary: Get Movie.
      tags:
      - Movie
    patch:
      description: Update movie by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: movie id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an movie
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.movieInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
      security:
      - BearerToken: []
      summary: Update Movie.
      tags:
      - Movie
  /register:
    post:
      description: registering a user from public access.
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user.
      tags:
      - Auth
swagger: "2.0"